[UndockingFix] Fixed undocked windows issue when toggling UI visible/hidden
Resolved using this git issue as reference: https://github.com/ocornut/imgui/issues/4757
Basically, don't call ImGui commands or at least Begin() if the docking context wasn't already submitted/Begin() this frame.
Also changing s_FramesCompleted from u64 to u32, cause like, that's kinda overkill as is.
Also moved notification code to the appropriate window, and stopped rendering camera test UI as it is still drawn then UI is hidden at the moment, albeit behind the scene.

+++ Source/Editor/QE_Editor.h
@@ -13,7 +13,7 @@ namespace QwerkE {
 
 		class EditorWindow;
 
-		typedef unsigned int u32;
+		typedef unsigned int u32; // 0 to 4,294,967,295
 		enum EditorStateFlags : u32
 		{
 			EditorStateFlagsNone			= 0,
@@ -40,7 +40,7 @@ namespace QwerkE {
 
 		bool ShowingEditorUI();
 
-		u64 CurrentFrame();
+		u32 CurrentFrame();
 
 		// Editor windows
 		void NewEditorWindow(u32 editorWindowType);
